backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
pdfplumber
python-docx
python-multipart
python-dotenv
